# Schems/swagger.yaml
openapi: 3.0.4
info:
  title: OpenVPN Gate Monitor API
  version: v1
paths:
...
  /api/open-vpn-clients/get-all-connected:
    get:
      tags:
        - OpenVpnServerClients
      parameters:
        - name: VpnServerId
          in: query
          required: true
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
        - name: PageSize
          in: query
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ConnectedClientsResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedClientsResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ConnectedClientsResponseApiResponse'
  /api/open-vpn-clients/get-all-history:
    get:
      tags:
        - OpenVpnServerClients
      parameters:
        - name: VpnServerId
          in: query
          required: true
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
        - name: PageSize
          in: query
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ConnectedClientsResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedClientsResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ConnectedClientsResponseApiResponse'
  /api/open-vpn-clients/overview/series:
    get:
      tags:
        - OpenVpnServerClients
      parameters:
        - name: From
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: To
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: Grouping
          in: query
          schema:
            $ref: '#/components/schemas/OverviewGrouping'
        - name: VpnServerId
          in: query
          schema:
            type: integer
            format: int32
        - name: ExternalId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OverviewSeriesResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewSeriesResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OverviewSeriesResponseApiResponse'
  /api/open-vpn-clients/overview/summary:
    get:
      tags:
        - OpenVpnServerClients
      parameters:
        - name: From
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: To
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: VpnServerId
          in: query
          schema:
            type: integer
            format: int32
        - name: ExternalId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OverviewTotalsResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewTotalsResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OverviewTotalsResponseApiResponse'
  /api/open-vpn-clients/overview/points:
    get:
      tags:
        - OpenVpnServerClients
      parameters:
        - name: From
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: To
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: VpnServerId
          in: query
          schema:
            type: integer
            format: int32
        - name: ExternalId
          in: query
          schema:
            type: string
        - name: OnlyWithCoordinates
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OverviewPointsResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewPointsResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OverviewPointsResponseApiResponse'
  /api/open-vpn-clients/overview/users:
    get:
      tags:
        - OpenVpnServerClients
      parameters:
        - name: From
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: To
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: VpnServerId
          in: query
          schema:
            type: integer
            format: int32
        - name: ExternalId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OverviewUsersResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewUsersResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OverviewUsersResponseApiResponse'
...
components:
  schemas:
...
    ConnectedClientsResponse:
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
        clients:
          type: array
          items:
            $ref: '#/components/schemas/VpnClientInfoResponse'
          nullable: true
      additionalProperties: false
    ConnectedClientsResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/ConnectedClientsResponse'
      additionalProperties: false
...
    GeoPointAggDto:
      type: object
      properties:
        country:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        sessionsCount:
          type: integer
          format: int32
        totalBytesIn:
          type: integer
          format: int64
        totalBytesOut:
          type: integer
          format: int64
      additionalProperties: false
...
    OverviewGrouping:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
      type: integer
      format: int32
    OverviewMetaDto:
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        grouping:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        trafficUnit:
          type: string
          nullable: true
        vpnServerId:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    OverviewPointsResponse:
      type: object
      properties:
        geoPointAgg:
          type: array
          items:
            $ref: '#/components/schemas/GeoPointAggDto'
          nullable: true
      additionalProperties: false
    OverviewPointsResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/OverviewPointsResponse'
      additionalProperties: false
    OverviewSeriesResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/OverviewMetaDto'
        summary:
          $ref: '#/components/schemas/OverviewSummaryDto'
        series:
          type: array
          items:
            $ref: '#/components/schemas/OverviewSeriesRowDto'
          nullable: true
      additionalProperties: false
    OverviewSeriesResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/OverviewSeriesResponse'
      additionalProperties: false
    OverviewSeriesRowDto:
      type: object
      properties:
        ts:
          type: string
          format: date-time
        activeClients:
          type: integer
          format: int32
        trafficInBytes:
          type: integer
          format: int64
        trafficOutBytes:
          type: integer
          format: int64
        trafficTotalBytes:
          type: integer
          format: int64
      additionalProperties: false
    OverviewSummaryDto:
      type: object
      properties:
        totalTrafficInBytes:
          type: integer
          format: int64
        totalTrafficOutBytes:
          type: integer
          format: int64
        peakActiveClients:
          type: integer
          format: int32
      additionalProperties: false
    OverviewTotalsResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/OverviewMetaDto'
        totals:
          $ref: '#/components/schemas/TotalsPayloadDto'
      additionalProperties: false
    OverviewTotalsResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/OverviewTotalsResponse'
      additionalProperties: false
    OverviewUserItem:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        vpnServerId:
          type: integer
          format: int32
          nullable: true
        sessions:
          type: integer
          format: int32
        trafficInBytes:
          type: integer
          format: int64
        trafficOutBytes:
          type: integer
          format: int64
        trafficTotalBytes:
          type: integer
          format: int64
          readOnly: true
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
      additionalProperties: false
    OverviewUsersResponse:
      type: object
      properties:
        overviewUserItems:
          type: array
          items:
            $ref: '#/components/schemas/OverviewUserItem'
          nullable: true
      additionalProperties: false
    OverviewUsersResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/OverviewUsersResponse'
      additionalProperties: false
...
    TotalsPayloadDto:
      type: object
      properties:
        sessionsCount:
          type: integer
          format: int64
        usersCount:
          type: integer
          format: int64
        accountsCount:
          type: integer
          format: int64
        trafficInBytes:
          type: integer
          format: int64
        trafficOutBytes:
          type: integer
          format: int64
        trafficTotalBytes:
          type: integer
          format: int64
          readOnly: true
      additionalProperties: false
...
    VpnClientInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        vpnServerId:
          type: integer
          format: int32
        externalId:
          type: string
          nullable: true
        tgUsername:
          type: string
          nullable: true
        tgFirstName:
          type: string
          nullable: true
        tgLastName:
          type: string
          nullable: true
        sessionId:
          type: string
          format: uuid
        commonName:
          type: string
          nullable: true
        remoteIp:
          type: string
          nullable: true
        localIp:
          type: string
          nullable: true
        bytesReceived:
          type: integer
          format: int64
        bytesSent:
          type: integer
          format: int64
        connectedSince:
          type: string
          format: date-time
        username:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        isConnected:
          type: boolean
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Enter the token in the format ''Bearer {your_token}'''
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]